requirements.txt
# ========= 基础依赖 =========
numpy>=1.23
pandas>=2.0
pyyaml>=6.0
tqdm>=4.65
joblib>=1.3
scikit-learn>=1.2

# ========= 数据 & 格式 =========
pyarrow>=14.0
fastparquet>=2023.1
tables>=3.8.0

# ========= 股票数据源 =========
baostock>=0.8.8
akshare>=1.11.80       # 如果你完全不用 akshare，可以删掉

# ========= 机器学习模型 =========
xgboost>=2.0.3          # GPU 支持核心版本
lightgbm>=4.1           # 可选，用于 future 强化版模型
catboost>=1.2.2         # 可选，用于 future 多模型融合

# ========= GPU 加速可选 =========
cupy-cuda11x>=12.0      # 若你的 CUDA = 11.x（如 11.2、11.8 等）
# cupy-cuda12x>=12.0    # 若你的 CUDA = 12.x（你按自己机器版本来）


# ========= 可选科学计算 =========
scipy>=1.10
matplotlib>=3.7
seaborn>=0.12

# ========= 你的项目本地模块 =========
# （项目自身的 src/ 下代码无需 pip 安装）
